refactor:table,order의 dao를 repository로 대체
