+++ src/main/java/kitchenpos/application/TableService.java
@@ -1,14 +1,15 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
-import kitchenpos.domain.OrderTableValidator;
 import kitchenpos.domain.order.OrderRepository;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.OrderTableRepository;
 import kitchenpos.ui.dto.OrderTableDto;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+
+import static kitchenpos.domain.table.OrderTableValidator.validateChangeEmptyTableOrderCondition;
 
 @Service
 public class TableService {
@@ -36,7 +37,7 @@ public class TableService {
     public OrderTable changeEmpty(final Long orderTableId, final OrderTableDto orderTableDto) {
         final OrderTable orderTable = orderTableRepository.findById(orderTableId)
                 .orElseThrow(() -> new IllegalArgumentException("테이블이 존재하지 않습니다. 빈 테이블로 변경할 수 없습니다."));
-        OrderTableValidator.validateChangeEmpty(orderTable, orderRepository.findByOrderTable(orderTable));
+        validateChangeEmptyTableOrderCondition(orderRepository.findByOrderTable(orderTable));
 
         orderTable.changeEmpty(orderTableDto.isEmpty());
