+++ src/main/java/kitchenpos/domain/order/OrderRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.order;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
