+++ src/main/java/kitchenpos/domain/table/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
 import org.springframework.data.jpa.repository.JpaRepository;
